update gh actions workflow traces to metrics
